+++ packages/components/src/Shared/Modal/ModalByEntityDomain/Climate/index.tsx
@@ -1,6 +1,4 @@
-import { ClimateControls } from "@components";
-import type { ClimateControlsProps } from "@components";
-
+import { ClimateControls, ClimateControlsProps } from "../../../Entity/Climate/ClimateControls";
 export type ModalClimateControlsProps = Omit<ClimateControlsProps, "style">;
 
 export function ModalClimateControls(props: ModalClimateControlsProps) {