+++ packages/components/src/Shared/Modal/ModalByEntityDomain/MediaPlayer/index.tsx
@@ -1,5 +1,5 @@
-import { MediaPlayerControls, type MediaPlayerControlsProps } from "@components";
 import type { EntityName, FilterByDomain } from "@hakit/core";
+import { MediaPlayerControls, MediaPlayerControlsProps } from "../../../Entity/MediaPlayer/MediaPlayerControls";
 export interface ModalMediaPlayerControlsProps extends MediaPlayerControlsProps {
   entity: FilterByDomain<EntityName, "media_player">;
 }